Move .filterable() to its own trait and update docs tests
