+++ provider/core/src/filter/impls.rs
@@ -6,77 +6,95 @@ use super::*;
 use crate::prelude::*;
 use icu_locid::LanguageIdentifier;
 
-/*
-/// Filter a data provider based on the language identifier in a data request.
-///
-/// Import this trait to add the methods on it to all DataProviders:
-///
-/// ```
-/// use icu_provider::filter::LanguageIdentifierFilter;
-///
-/// // now you can call `.filter_by_langid()` on any provider
-/// ```
-///
-/// # Examples
-///
-/// ```
-/// use icu_provider::prelude::*;
-/// use icu_provider::hello_world::*;
-/// use icu_provider::filter::LanguageIdentifierFilter;
-/// use icu_provider::iter::IterableDataProviderCore;
-/// use icu_locid::LanguageIdentifier;
-/// use icu_locid_macros::{language, langid};
-///
-/// let provider = HelloWorldProvider::new_with_placeholder_data()
-///     .filter_by_langid(|langid| langid.language != language!("en"));
-///
-/// // German requests should succeed:
-/// let req_de = DataRequest {
-///     resource_path: ResourcePath {
-///         key: key::HELLO_WORLD_V1,
-///         options: langid!("de").into(),
-///     }
-/// };
-/// let response: Result<DataResponse<HelloWorldV1Marker>, _> =
-///     provider.load_payload(&req_de);
-/// assert!(matches!(response, Ok(_)));
-///
-/// // English requests should fail:
-/// let req_en = DataRequest {
-///     resource_path: ResourcePath {
-///         key: key::HELLO_WORLD_V1,
-///         options: langid!("en-US").into(),
-///     }
-/// };
-/// let response: Result<DataResponse<HelloWorldV1Marker>, _> =
-///     provider.load_payload(&req_en);
-/// assert!(matches!(response, Err(DataError::FilteredResource(_, _))));
-///
-/// // English should not appear in the iterator result:
-/// let supported_langids = provider.supported_options_for_key(&key::HELLO_WORLD_V1)
-///     .expect("Should successfully make an iterator of supported locales")
-///     .filter_map(|options| options.langid)
-///     .collect::<Vec<LanguageIdentifier>>();
-/// assert!(supported_langids.contains(&langid!("de")));
-/// assert!(!supported_langids.contains(&langid!("en")));
-/// ```
-pub trait LanguageIdentifierFilter<M>: Sized {
+impl<D, F> RequestFilterDataProvider<D, F>
+where
+    F: Fn(&DataRequest) -> bool,
+{
     /// Filter out data requests with certain langids according to the predicate function. The
     /// predicate should return `true` to allow a langid and `false` to reject a langid.
     ///
     /// Data requests with no langid will be allowed. To reject data requests without a langid,
     /// chain this with [`Self::require_langid`].
-    fn filter_by_langid<'a>(
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use icu_provider::prelude::*;
+    /// use icu_provider::hello_world::*;
+    /// use icu_provider::filter::Filterable;
+    /// use icu_provider::iter::IterableDataProviderCore;
+    /// use icu_locid::LanguageIdentifier;
+    /// use icu_locid_macros::{language, langid};
+    ///
+    /// let provider = HelloWorldProvider::new_with_placeholder_data()
+    ///     .filterable()
+    ///     .filter_by_langid(|langid| langid.language != language!("en"));
+    ///
+    /// // German requests should succeed:
+    /// let req_de = DataRequest {
+    ///     resource_path: ResourcePath {
+    ///         key: key::HELLO_WORLD_V1,
+    ///         options: langid!("de").into(),
+    ///     }
+    /// };
+    /// let response: Result<DataResponse<HelloWorldV1Marker>, _> =
+    ///     provider.load_payload(&req_de);
+    /// assert!(matches!(response, Ok(_)));
+    ///
+    /// // English requests should fail:
+    /// let req_en = DataRequest {
+    ///     resource_path: ResourcePath {
+    ///         key: key::HELLO_WORLD_V1,
+    ///         options: langid!("en-US").into(),
+    ///     }
+    /// };
+    /// let response: Result<DataResponse<HelloWorldV1Marker>, _> =
+    ///     provider.load_payload(&req_en);
+    /// assert!(matches!(response, Err(DataError::FilteredResource(_, _))));
+    ///
+    /// // English should not appear in the iterator result:
+    /// let supported_langids = provider.supported_options_for_key(&key::HELLO_WORLD_V1)
+    ///     .expect("Should successfully make an iterator of supported locales")
+    ///     .filter_map(|options| options.langid)
+    ///     .collect::<Vec<LanguageIdentifier>>();
+    /// assert!(supported_langids.contains(&langid!("de")));
+    /// assert!(!supported_langids.contains(&langid!("en")));
+    /// ```
+    pub fn filter_by_langid<'a>(
         self,
         predicate: impl Fn(&LanguageIdentifier) -> bool + 'a,
-    ) -> RequestFilterDataProvider<Self, Box<dyn Fn(&DataRequest) -> bool + 'a>>;
+    ) -> RequestFilterDataProvider<D, Box<dyn Fn(&DataRequest) -> bool + 'a>>
+    where
+        F: 'a,
+    {
+        self.filter_by_langid_with_description(predicate, "Locale filter".to_string())
+    }
 
-    /// Like `Box<dyn Fn(&DataRequest) -> bool + 'a>ilter_by_langid`, but with a string description to help with debugging.
-    fn filter_by_langid_with_description<'a>(
+    /// Same as [`Self::filter_by_langid`] but with an extra argument to set a custom
+    /// description for debugging.
+    pub fn filter_by_langid_with_description<'a>(
         self,
         predicate: impl Fn(&LanguageIdentifier) -> bool + 'a,
         description: String,
-    ) -> RequestFilterDataProvider<Self, Box<dyn Fn(&DataRequest) -> bool + 'a>>;
+    ) -> RequestFilterDataProvider<D, Box<dyn Fn(&DataRequest) -> bool + 'a>>
+    where
+        F: 'a,
+    {
+        let old_predicate = self.predicate;
+        RequestFilterDataProvider {
+            inner: self.inner,
+            predicate: Box::new(move |request| -> bool {
+                if !(old_predicate)(request) {
+                    return false;
+                }
+                match &request.resource_path.options.langid {
+                    Some(langid) => predicate(langid),
+                    None => true,
+                }
+            }),
+            description,
+        }
+    }
 
     /// Filter out data request except those having a language identifier that exactly matches
     /// one in the allowlist.
@@ -92,11 +110,12 @@ pub trait LanguageIdentifierFilter<M>: Sized {
     /// ```
     /// use icu_provider::prelude::*;
     /// use icu_provider::hello_world::*;
-    /// use icu_provider::filter::LanguageIdentifierFilter;
+    /// use icu_provider::filter::Filterable;
     /// use icu_locid_macros::langid;
     ///
     /// let allowlist = vec![langid!("de"), langid!("zh")];
     /// let provider = HelloWorldProvider::new_with_placeholder_data()
+    ///     .filterable()
     ///     .filter_by_langid_allowlist_strict(&allowlist);
     ///
     /// // German requests should succeed:
@@ -125,55 +144,6 @@ pub trait LanguageIdentifierFilter<M>: Sized {
     ///     response.unwrap_err().to_string()
     /// );
     /// ```
-    fn filter_by_langid_allowlist_strict<'a>(
-        self,
-        allowlist: &'a [LanguageIdentifier],
-    ) -> RequestFilterDataProvider<Self, Box<dyn Fn(&DataRequest) -> bool + 'a>>;
-
-    fn require_langid<'a>(
-        self,
-    ) -> RequestFilterDataProvider<Self, Box<dyn Fn(&DataRequest) -> bool + 'a>>;
-}
-*/
-
-impl<D, F> RequestFilterDataProvider<D, F>
-where
-    F: Fn(&DataRequest) -> bool,
-{
-    pub fn filter_by_langid<'a>(
-        self,
-        predicate: impl Fn(&LanguageIdentifier) -> bool + 'a,
-    ) -> RequestFilterDataProvider<D, Box<dyn Fn(&DataRequest) -> bool + 'a>>
-    where
-        F: 'a,
-    {
-        self.filter_by_langid_with_description(predicate, "Locale filter".to_string())
-    }
-
-    pub fn filter_by_langid_with_description<'a>(
-        self,
-        predicate: impl Fn(&LanguageIdentifier) -> bool + 'a,
-        description: String,
-    ) -> RequestFilterDataProvider<D, Box<dyn Fn(&DataRequest) -> bool + 'a>>
-    where
-        F: 'a,
-    {
-        let old_predicate = self.predicate;
-        RequestFilterDataProvider {
-            inner: self.inner,
-            predicate: Box::new(move |request| -> bool {
-                if !(old_predicate)(request) {
-                    return false;
-                }
-                match &request.resource_path.options.langid {
-                    Some(langid) => predicate(langid),
-                    None => true,
-                }
-            }),
-            description,
-        }
-    }
-
     pub fn filter_by_langid_allowlist_strict<'a>(
         self,
         allowlist: &'a [LanguageIdentifier],
@@ -204,6 +174,7 @@ where
     /// ```
     /// use icu_provider::prelude::*;
     /// use icu_provider::hello_world::*;
+    /// use icu_provider::filter::Filterable;
     /// use icu_locid_macros::langid;
     ///
     /// let provider = HelloWorldProvider::new_with_placeholder_data()
